Implement Field::HasMany with a table on show page
https://trello.com/c/NygJOCUM
https://trello.com/c/TERo00zt

On the `index` page, `has_many` relationships are displayed by simple
text telling how many associated objects there are.

On the `show` page, `has_many` relationships are displayed by an HTML
table, identical to the one that appears on the associated object's
`index` page.

On the `form` pages, `has_many` relationships show a message saying the
functionality is not yet implemented.

- Extracted the common table HTMl into a partial
- Extract strings to I18n
- Rename index_page_attributes -> table_attributes
- Rename `Page::Index` -> `Page::Table`
    We're now using the class to display tables both on the index page
    and on the other resource's show pages through the HasMany adapter.

Follow-up tasks:

- Implement forms for `has_many` relationships
  (needs discussion: https://trello.com/c/gjsGaU5W).