Generate a single controller for all resources
Why?

When we ran the generators on Hound, there was a lot of feedback that
people didn't want the extra generated controller files. We didn't
need to customize the controllers, and we ended up changing the
routing so all dashboards were served by a single controller.

This change makes a single controller the default. As demonstrated with
the still-present `Admin::CustomersController`, users can still override
the controllers themselves.

We left the `Admin::CustomersController` in place because there are
tests that exercise `Administrate::ApplicationController` through it.

- Update changelog
- Update route generator
- Update specs to actually read generated routes file