Add sorting by index table columns
Problem:

A common use case for admin dashboards is sorting by a column,
and it feels restrictive to not have access to controls
for sorting the data.

Solution:

- Add column-based ordering

Users can click on a column heading to order by that column,
and they can click a second time to reverse the order.

The column that is currently being ordered by
displays an arrow representing the direction the data is ordered in.

Sorting does not interfere with the user's current search

Most of the ordering logic lives in `Administrate::Order`,
including knowledge of the currently ordered attribute
and the HTTP parameters for each sorting link.

If a user tries to order by any column that doesn't exist on the model,
it will quietly ignore the request and not sort the data at all.