Update README
Problem:

The README was written when the repo was simply a prototype app for
Administrate, before all of the application code was extracted to the
`adminsitrate` directory and released as its own gem.

When visitors to the github repo saw
"this is a demo app for Administrate",
it caused confusion and led to the assumption that the main Administrate
codebase lived somewhere else.

Solution:

Update README with the overview and some "getting started" snippets
from http://administrate-docs.herokuapp.com

Add a section describing repo organization.