Fix alignment issue for labels on show page
Fixes #230

Problem:

When an attribute's name is long enough
to wrap onto a second line on the show page,
and when the corresponding attribute-data element is empty,
following attribute labels don't clear to a new line.

This breaks the page layout for all attributes
following the offending one.

Fix:

Add `clear: left` to the `.attribute-label` selector,
to make sure it aligns correctly without interference from other labels.