Raise a helpful error for missing dashboard attrs
Closes #246

Problem:

If an attribute is missing from a dashboard's `ATTRIBUTE_TYPES` hash,
the user will see a generic error
when they visit a dashboard page that tries to display that attribute.

The error is:

```
NoMethodError: undefined method new for `nil:NilClass` in `Page::Base`
```

Solution:

Add accessor methods to `BaseDashboard`
for fetching attributes from the dashboard.
If a requested attribute is not present,
the accessor methods raise a helpful error message.